requirements.txt
flask
gunicorn
google-auth
google-auth-oauthlib
google-api-python-client
google-genai
requests
apscheduler
pandas
google-cloud-secret-manager
google-cloud-storage
google-cloud-datastore
google-cloud-tasks
